chore: renamed package to mony.api
