+++ API/src/main/java/io/github/invvk/mony/api/config/properties/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Properties package for all of Mony configs.
+ * @see io.github.invvk.mony.api.config.properties.ConfigProperty
+ * @see io.github.invvk.mony.api.config.properties.MessagesProperty
+ */
+package io.github.invvk.mony.api.config.properties;
