+++ API/src/main/java/io/github/invvk/mony/api/database/User.java
@@ -1,4 +1,4 @@
-package io.github.invvk.mony.database;
+package io.github.invvk.mony.api.database;
 
 import com.google.gson.annotations.Expose;
 import lombok.Data;
