+++ API/src/main/java/io/github/invvk/mony/api/database/misc/StorageMode.java
@@ -1,0 +1,7 @@
+package io.github.invvk.mony.api.database.misc;
+
+public enum StorageMode {
+
+    FILE, MYSQL
+
+}
