--- API/src/main/java/io/github/invvk/mony/config/properties/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Properties package for all of Mony configs.
- * @see io.github.invvk.mony.config.properties.ConfigProperty
- * @see io.github.invvk.mony.config.properties.MessagesProperty
- */
-package io.github.invvk.mony.config.properties;
