--- API/src/main/java/io/github/invvk/mony/database/misc/StorageMode.java
@@ -1,7 +1,0 @@
-package io.github.invvk.mony.database.misc;
-
-public enum StorageMode {
-
-    FILE, MYSQL
-
-}
