--- API/src/main/java/io/github/invvk/mony/database/storage/IStorage.java
@@ -1,13 +1,0 @@
-package io.github.invvk.mony.database.storage;
-
-import io.github.invvk.mony.database.manager.IDataManager;
-
-public interface IStorage {
-
-    void init();
-
-    void close();
-
-    IDataManager getDataManager();
-
-}
