+++ Core/src/main/java/io/github/invvk/mony/internal/MonyBootstrap.java
@@ -2,11 +2,11 @@ package io.github.invvk.mony.internal;
 
 import io.github.invvk.mony.internal.hook.placeholderapi.PlaceholderHook;
 import io.github.invvk.mony.internal.listener.DailyLimitListener;
-import io.github.invvk.mony.Mony;
-import io.github.invvk.mony.config.ConfigManager;
+import io.github.invvk.mony.api.Mony;
+import io.github.invvk.mony.api.config.ConfigManager;
 import io.github.invvk.mony.internal.database.StorageManager;
 import io.github.invvk.mony.internal.database.UserManager;
-import io.github.invvk.mony.database.storage.IStorage;
+import io.github.invvk.mony.api.database.storage.IStorage;
 import io.github.invvk.mony.internal.hook.VaultHook;
 import io.github.invvk.mony.internal.listener.ConnectionListener;
 import io.github.invvk.mony.internal.listener.MobKillListener;
