+++ Core/src/main/java/io/github/invvk/mony/internal/database/manager/FileDataManager.java
@@ -2,8 +2,8 @@ package io.github.invvk.mony.internal.database.manager;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import io.github.invvk.mony.database.User;
-import io.github.invvk.mony.database.manager.IDataManager;
+import io.github.invvk.mony.api.database.User;
+import io.github.invvk.mony.api.database.manager.IDataManager;
 import lombok.RequiredArgsConstructor;
 
 import java.io.File;
