+++ Core/src/main/java/io/github/invvk/mony/internal/database/storage/MySQLStorage.java
@@ -5,9 +5,9 @@ import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
 import io.github.invvk.mony.internal.MonyBootstrap;
 import io.github.invvk.mony.internal.database.manager.MySQLDataManager;
-import io.github.invvk.mony.config.properties.ConfigProperty;
-import io.github.invvk.mony.database.manager.IDataManager;
-import io.github.invvk.mony.database.storage.IStorage;
+import io.github.invvk.mony.api.config.properties.ConfigProperty;
+import io.github.invvk.mony.api.database.manager.IDataManager;
+import io.github.invvk.mony.api.database.storage.IStorage;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
