+++ Core/src/main/java/io/github/invvk/mony/internal/listener/ConnectionListener.java
@@ -1,9 +1,9 @@
 package io.github.invvk.mony.internal.listener;
 
 import io.github.invvk.mony.internal.MonyBootstrap;
-import io.github.invvk.mony.config.properties.ConfigProperty;
-import io.github.invvk.mony.database.User;
-import io.github.invvk.mony.events.MonyCacheLoadEvent;
+import io.github.invvk.mony.api.config.properties.ConfigProperty;
+import io.github.invvk.mony.api.database.User;
+import io.github.invvk.mony.api.events.MonyCacheLoadEvent;
 import lombok.RequiredArgsConstructor;
 import org.bukkit.Bukkit;
 import org.bukkit.event.EventHandler;
