+++ Core/src/test/java/io/github/invvk/mony/test/user/TestUserManager.java
@@ -3,7 +3,7 @@ package io.github.invvk.mony.test.user;
 import be.seeseemelk.mockbukkit.MockBukkit;
 import be.seeseemelk.mockbukkit.ServerMock;
 import io.github.invvk.mony.internal.MonyLoader;
-import io.github.invvk.mony.database.User;
+import io.github.invvk.mony.api.database.User;
 import org.junit.jupiter.api.*;
 
 import java.util.UUID;
